Backport to .NET5 (#15)
* Backport library to .NET5

The decision to simply bump the version to .NET7 was purely out of convenience because it provided a tiny little utility function that the Matrix implementation uses. This decision was made when this wasn't really intended to be a proper continuation of the project but simply a quick and dirty fork that would work for my purposes. I realize that considering the broad userbase people might genuinely benefit from this.

* Oops, CI needs .NET6

* Use `Array.Empty` instead of collection expressions

Linux does not like them unfortunately

* Missed one

* Fix warning